Split into BepInEx and MonoGame projects
BepInEx mod works as usual. I was able to get the mod to launch in the MonoGame port on the modding_support branch with a couple of changes: 1) loading all of the assemblies before looking through them to find IMod types, and 2) replacing all calls to GetExecutingAssembly with GetEntryAssembly, as HarmonyX/MonoMod replaces the executing assembly.